The boiler plate was having babel issues. I moved actions, reducers, and other relevant files into the overarching application folder. I also commented out the router portion of the the index.js so that the component would render.
