+++ .babelrc
@@ -1,0 +1,3 @@
+{
+  "presets": ["es2015"] 
+}
