+++ app/components/apiHelper.js
@@ -1,5 +1,4 @@
 export default class ApiCall {
-
   static getAllData() {
     return fetch('https://thecountedapi.com/api/counted')
     .then(response => response.json())
