+++ app/index.js
@@ -7,22 +7,22 @@ import { ConnectedRouter, routerMiddleware } from 'react-router-redux';
 import createHistory from 'history/createBrowserHistory';
 // import { Route } from 'react-router-dom';
 
-import { rootReducer } from '../src/reducers/index.js';
-import { Main } from '../src/components/Main';
-import { addData } from '../src/actions/index.js';
+import rootReducer from './reducers/index.js';
+
+import  Main from './components/Main';
+import { loadData } from './actions/index.js';
 
 const history = createHistory();
 const routeMiddleware = routerMiddleware(history);
 const devTools = window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__();
 const store = createStore(rootReducer, devTools, applyMiddleware(thunk, routeMiddleware));
 
-store.dispatch(addData());
-
+// store.dispatch(loadData());
 
 ReactDOM.render(
   <Provider store={store}>
-    <ConnectedRouter history={history}>
+    {/* <ConnectedRouter history={history}> */}
       <Main />
-    </ConnectedRouter>
+    {/* </ConnectedRouter> */}
   </Provider>,
   document.getElementById('main'));
