+++ package.json
@@ -17,6 +17,7 @@
     "express": "^4.14.0",
     "express-cors": "0.0.3",
     "history": "^4.6.1",
+    "jsx-loader": "^0.13.2",
     "nodemon": "^1.11.0",
     "react": "^15.4.2",
     "react-dom": "^15.4.2",
@@ -29,18 +30,19 @@
     "redux-thunk": "^2.2.0"
   },
   "devDependencies": {
-    "babel-cli": "^6.18.0",
+    "babel-cli": "^6.24.1",
     "babel-core": "^6.21.0",
     "babel-jest": "^20.0.3",
     "babel-loader": "^6.2.10",
-    "babel-preset-es2015": "^6.18.0",
+    "babel-preset-es2015": "^6.24.1",
     "babel-preset-react": "^6.16.0",
+    "css-loader": "^0.28.4",
+    "enzyme": "^2.8.2",
     "jest": "^20.0.3",
+    "react-addons-test-utils": "^15.5.1",
+    "redux-mock-store": "^1.2.3",
     "webpack": "^1.14.0",
     "webpack-dev-middleware": "^1.9.0",
-    "webpack-hot-middleware": "^2.15.0",
-    "enzyme": "^2.8.2",
-    "react-addons-test-utils": "^15.5.1",
-    "redux-mock-store": "^1.2.3"
+    "webpack-hot-middleware": "^2.15.0"
   }
 }
